# docs/openapi/products-v3.yaml
openapi: 3.0.3
info:
  title: Mini E-Commerce Products API
  version: 3.0.0
  description: Contract-first Products CRUD API (v3) - Category required

servers:
  - url: /api/v3

tags:
  - name: Products
    description: Product management (v3)

paths:
  /products:
    get:
      tags: [Products]
      summary: List products (paginated, with category)
      operationId: listProductsV3
      parameters:
        - name: page
          in: query
          required: false
          description: Zero-based page index
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          description: Page size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: sort
          in: query
          required: false
          description: Sort format `field,asc|desc` (e.g. `createdAt,desc`)
          schema:
            type: string
            example: createdAt,desc
        - name: q
          in: query
          required: false
          description: Optional text search query (name/description)
          schema:
            type: string
            example: iphone
        - name: categoryId
          in: query
          required: false
          description: Filter by category ID
          schema:
            type: string
            example: "1"
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPageResponseV3'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Products]
      summary: Create a product (v3)
      operationId: createProductV3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductV3Request'
      responses:
        '201':
          description: Product created
          headers:
            Location:
              description: URL of the created resource
              schema:
                type: string
                example: /api/v3/products/prd_123
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseV3'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict (e.g., duplicate SKU if enforced)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /products/{id}:
    get:
      tags: [Products]
      summary: Get product by id (v3)
      operationId: getProductByIdV3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: prd_123
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseV3'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [Products]
      summary: Replace product (full update) (v3)
      operationId: replaceProductV3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: prd_123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductV3Request'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseV3'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product or category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict (e.g., duplicate SKU if enforced)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags: [Products]
      summary: Update product (partial update) (v3)
      operationId: patchProductV3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: prd_123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchProductV3Request'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseV3'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product or category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict (e.g., duplicate SKU if enforced)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Products]
      summary: Delete product (v3)
      operationId: deleteProductV3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: prd_123
      responses:
        '204':
          description: Product deleted
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:

    # ---------- Core Models ----------
    CategoryRef:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          example: "1"
        name:
          type: string
          example: Smartphones

    ProductV3:
      type: object
      required:
        - id
        - name
        - price
        - currency
        - inStock
        - category
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          example: prd_123
        sku:
          type: string
          description: Optional SKU (can be used for uniqueness if you want)
          example: SKU-IPH15-BLK
          nullable: true
        name:
          type: string
          minLength: 3
          maxLength: 120
          example: iPhone 15
        description:
          type: string
          maxLength: 2000
          nullable: true
          example: Latest Apple iPhone
        price:
          type: number
          format: double
          minimum: 0
          example: 999.99
        currency:
          type: string
          description: ISO 4217 currency code
          minLength: 3
          maxLength: 3
          example: USD
        inStock:
          type: boolean
          example: true
        imageUrl:
          type: string
          nullable: true
          example: https://cdn.example.com/products/prd_123.jpg
        discountPercent:
          type: number
          format: double
          minimum: 0
          maximum: 100
          nullable: true
          example: 10.0
        rating:
          type: number
          format: double
          minimum: 0
          maximum: 5
          nullable: true
          example: 4.5
        category:
          $ref: '#/components/schemas/CategoryRef'
        createdAt:
          type: string
          format: date-time
          example: "2026-01-26T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-01-26T10:30:00Z"

    # ---------- Requests ----------
    CreateProductV3Request:
      type: object
      required:
        - name
        - price
        - currency
        - inStock
        - imageUrl
        - discountPercent
        - rating
        - categoryId
      properties:
        sku:
          type: string
          nullable: true
          example: SKU-IPH15-BLK
        name:
          type: string
          minLength: 3
          maxLength: 120
          example: iPhone 15
        description:
          type: string
          maxLength: 2000
          nullable: true
          example: Latest Apple iPhone
        price:
          type: number
          format: double
          minimum: 0
          example: 999.99
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: USD
        inStock:
          type: boolean
          example: true
        imageUrl:
          type: string
          example: https://cdn.example.com/products/prd_123.jpg
        discountPercent:
          type: number
          format: double
          minimum: 0
          maximum: 100
          example: 10.0
        rating:
          type: number
          format: double
          minimum: 0
          maximum: 5
          example: 4.5
        categoryId:
          type: string
          description: Category ID (required)
          example: "1"

    UpdateProductV3Request:
      type: object
      required:
        - name
        - price
        - currency
        - inStock
        - imageUrl
        - discountPercent
        - rating
        - categoryId
      properties:
        sku:
          type: string
          nullable: true
          example: SKU-IPH15-BLK
        name:
          type: string
          minLength: 3
          maxLength: 120
          example: iPhone 15
        description:
          type: string
          maxLength: 2000
          nullable: true
          example: Latest Apple iPhone
        price:
          type: number
          format: double
          minimum: 0
          example: 999.99
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: USD
        inStock:
          type: boolean
          example: true
        imageUrl:
          type: string
          example: https://cdn.example.com/products/prd_123.jpg
        discountPercent:
          type: number
          format: double
          minimum: 0
          maximum: 100
          example: 10.0
        rating:
          type: number
          format: double
          minimum: 0
          maximum: 5
          example: 4.5
        categoryId:
          type: string
          description: Category ID (required)
          example: "1"

    PatchProductV3Request:
      type: object
      description: All fields optional; only provided fields will be updated
      properties:
        sku:
          type: string
          nullable: true
          example: SKU-IPH15-BLK
        name:
          type: string
          minLength: 3
          maxLength: 120
          nullable: true
          example: iPhone 15
        description:
          type: string
          maxLength: 2000
          nullable: true
          example: Latest Apple iPhone
        price:
          type: number
          format: double
          minimum: 0
          nullable: true
          example: 999.99
        currency:
          type: string
          minLength: 3
          maxLength: 3
          nullable: true
          example: USD
        inStock:
          type: boolean
          nullable: true
          example: true
        imageUrl:
          type: string
          nullable: true
          example: https://cdn.example.com/products/prd_123.jpg
        discountPercent:
          type: number
          format: double
          minimum: 0
          maximum: 100
          nullable: true
          example: 10.0
        rating:
          type: number
          format: double
          minimum: 0
          maximum: 5
          nullable: true
          example: 4.5
        categoryId:
          type: string
          nullable: true
          example: "1"
      additionalProperties: false

    # ---------- Responses ----------
    ProductResponseV3:
      type: object
      required: [product]
      properties:
        product:
          $ref: '#/components/schemas/ProductV3'

    ProductPageResponseV3:
      type: object
      required:
        - items
        - page
        - size
        - totalElements
        - totalPages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductV3'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 10
        totalElements:
          type: integer
          example: 120
        totalPages:
          type: integer
          example: 12

    # ---------- Error ----------
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Validation failed
        details:
          type: array
          description: Optional list of field-level errors
          items:
            $ref: '#/components/schemas/ErrorDetail'
          nullable: true
        traceId:
          type: string
          description: Optional correlation/trace id for debugging
          nullable: true
          example: 5f2c3d8a8b2b4c1c

    ErrorDetail:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          example: name
        message:
          type: string
          example: must be at least 3 characters